renamed main, section component
